+++ RELEASE_NOTES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Add support for AppInsufficientResourceError in executionPolicy.restartOn
+
 ## 2.15.0 2025-09-29
 
 * Added support for new region in OCI Ashburn